Updates for items docs and splitting of tables api
